refactor(web): decompose FinykApp.tsx and fizruk adapter, extract shared dualWrite framework
- FinykApp.tsx: 656→484 lines - extract SyncIndicator helper, inline components
- fizruk/adapter.ts: 642→102 lines - split into ops/workouts, ops/exercises, ops/dailyPlanTemplates
- Create shared/lib/dualWrite/core.ts with generic ApplyDualWriteOptions/DualWriteLogger types
- Fix react-hooks/exhaustive-deps in FinykApp.tsx (3 violations)
- Fix react-hooks/exhaustive-deps in useWorkoutsLifecycle.ts (2 violations)
- Update technical-assessment-2026-06-05.md with reduced eslint-disable count
- Update initiative 0021 with completed items
