+++ apps/server/eslint.server-maxlines-allowlist.json
@@ -1,0 +1,14 @@
+[
+  "apps/server/src/modules/chat/chat.ts",
+  "apps/server/src/http/rateLimit.ts",
+  "apps/server/src/modules/alerts/store.ts",
+  "apps/server/src/modules/billing/stripe.ts",
+  "apps/server/src/auth.ts",
+  "apps/server/src/lib/strategicGoals.ts",
+  "apps/server/src/modules/sync/fizruk/applySync.ts",
+  "apps/server/src/modules/ai-memory/ingestQueue.ts",
+  "apps/server/src/routes/internal/openclaw.ts",
+  "apps/server/src/env/env.ts",
+  "apps/server/src/modules/openclaw/tools.ts",
+  "apps/server/src/obs/metrics.ts"
+]
