+++ apps/web/src/modules/fizruk/lib/dualWrite/ops/index.ts
@@ -1,0 +1,24 @@
+// Re-export all operation handlers from subdirectory modules
+export {
+  upsertWorkout,
+  upsertWorkoutItem,
+  upsertWorkoutSet,
+  softDeleteWorkout,
+} from "./workouts";
+
+export {
+  upsertCustomExercise,
+  softDeleteCustomExercise,
+  upsertMeasurement,
+  softDeleteMeasurement,
+} from "./exercises";
+
+export {
+  upsertDailyLog,
+  softDeleteDailyLog,
+  setMonthlyPlan,
+  upsertWorkoutTemplate,
+  softDeleteWorkoutTemplate,
+} from "./dailyPlanTemplates";
+
+export type { WorkoutSet } from "./workouts";
