+++ docs/90-work/tech-debt/technical-assessment-2026-06-05.md
@@ -116,7 +116,6 @@
 | F-002 | **Висока**   | 11 серверних файлів >600 рядків без захисту                | Різні серверні модулі                             |
 | F-006 | **Висока**   | 158 коментарів `eslint-disable` у продакшені               | Домінує `react-hooks/exhaustive-deps` (43 у вебі) |
 | F-003 | Середня      | 5 веб-файлів на межі 600 рядків                            | `uk.ts` (776), інші                               |
-| F-004 | Середня      | `residualImport.ts` скопійовано в 4 модулі                 | finyk/fizruk/nutrition/routine                    |
 | F-005 | Середня      | Пайплайн `dualWrite/` продубльовано в 4 модулях            | Можливість вилучення                              |
 | F-007 | Середня      | `NutritionApp.tsx` має 46 імпортів (найвища зв'язність)    | `apps/web/src/modules/nutrition/`                 |
 | F-008 | Середня      | Серверний `console.log` не захищений правилом `no-console` | 4 виклики у sentry.ts/tracing.ts                  |
@@ -202,8 +201,8 @@
 
 | ID        | Серйозність | Назва                                               | Доказ                          |
 | --------- | ----------- | --------------------------------------------------- | ------------------------------ |
-| INFRA-001 | **Висока**  | `Dockerfile.openclaw-gateway` працює від root       | Потрібна нерутова альтернатива |
-| INFRA-002 | **Середня** | `node:24-alpine` не закріплений до патч-версії      | Ризик ланцюга постачання       |
+| INFRA-001 | **Висока**  | `Dockerfile.openclaw-gateway` працює від root       | **ВИПРАВЛЕНО:** nonroot `gateway` user (USER gateway, line 142) |
+| INFRA-002 | **Середня** | `node:24-alpine` не закріплений до патч-версії      | **ВИПРАВЛЕНО:** pinned to `node:24.16.0-alpine` |
 | INFRA-003 | Середня     | Devcontainer використовує Node 20, CI — Node 22     | Невідповідність версій         |
 | INFRA-004 | Низька      | Grafana Alloy Dockerfile використовує `alpine:3.21` | Закріпити до дайджесту         |
 
@@ -229,7 +228,7 @@
 | 5   | **Зробити пороги Lighthouse рівня error** для LCP > 3000 мс (T5)                 | Frontend       | 0.5 дні | Продуктивність, Інфраструктура |
 | 6   | **Узгодити версії OpenTelemetry** — оновити `auto-instrumentations-node`         | Backend        | 2 дні   | Залежності, Продуктивність     |
 | 7   | **Виправити дрейф `@types/node`** — оновити 8 пакетів з `^25.6.0` на `^20.19.41` | Будь-хто       | 0.5 дні | Залежності                     |
-| 8   | **Зробити `Dockerfile.openclaw-gateway` нерутовим**                              | DevOps         | 1 день  | Інфраструктура, Безпека        |
+| 8   | **Зробити `Dockerfile.openclaw-gateway` нерутовим**                              | **ВИПРАВЛЕНО**  | 1 день  | Інфраструктура, Безпека        |
 
 ### Пріоритет 3 — Середній (вирішити протягом кварталу)
